requirements.txt
PyYAML>=6.0
slack_sdk>=3.27
